Category Updated

+++ src/app/admin/category/category.component.ts
@@ -10,6 +10,7 @@ import { Category } from 'src/app/models/Category';
 })
 export class CategoryComponent implements OnInit {
 
+  defaultImageURL: any ="assets/images/bag1.png";
   imageURL:any="assets/images/bag1.png";
   filetoUpload:File;
   categories : Category[];
@@ -20,16 +21,20 @@ export class CategoryComponent implements OnInit {
     this.createForm();
     this.getCategoryList();
   }
+  get CategoryName(){   return this.categoryService.CategoryForm.get('CategoryName');  }
+  get Description(){   return this.categoryService.CategoryForm.get('Description');  }
+  get Image(){   return this.categoryService.CategoryForm.get('Image');  }
   createForm()
   {
     this.categoryService.CategoryForm=this.fb.group(
       {
-        CategoryID : [""],
+        CategoryID : [0],
         CategoryName : ["",Validators.required],
         Description : ["",Validators.required],
-        Image : [null,Validators.required]
+        Image : [null,[Validators.required]]
       });
   }
+
   onImageChange(event:any)
   {
     console.log(event);
@@ -46,21 +51,32 @@ export class CategoryComponent implements OnInit {
     }
     reader.readAsDataURL(this.filetoUpload);
   }
+
   onSubmit()
   {
     let category:Category=this.categoryService.CategoryForm.value;
-    this.categoryService.add(category).subscribe(
-      (data:any)=>{
-        console.log("Category added successfully!");
-        this.categories=data;
-        this.categoryService.categoryList=data;
-        this.createForm();
-      },
-      (err:any)=>{
-        console.log("Error occurred!");
-      }
-    );
+    console.log(category)
+    if(category.CategoryID!=0)
+    {
+        this.update(category);
+    }
+    else
+    {
+      this.categoryService.add(category).subscribe(
+        (data:any)=>{
+          console.log("Category added successfully!");
+          this.categories=data;
+          this.categoryService.categoryList=data;
+          this.resetForm();         
+        },
+        (err:any)=>{
+          console.log("Error occurred!");
+        }
+      );
+    }
+
   }
+
   getCategoryList()
   {
     this.categoryService.getAll().subscribe(
@@ -73,6 +89,7 @@ export class CategoryComponent implements OnInit {
       }
     );
   }
+
   search(event:any)
   {
     console.log(event.target.value);
@@ -92,16 +109,51 @@ export class CategoryComponent implements OnInit {
       this.categories=this.categoryService.categoryList;
     }
   }
+
   deleteCategory(category:Category)
   {
-    this.categoryService.delete(category.CategoryID).subscribe(
-      (data:any)=>{
+    if(confirm('Are you sure want to delete? ')){
+      this.categoryService.delete(category.CategoryID).subscribe(
+        (data:any)=>{
+          this.categories=data;
+          this.categoryService.categoryList=data;
+        },
+        (err:any)=>{
+          console.log("Error occurred!");
+        }
+      );
+    }
+
+  }
+
+  editCategory(category: Category)
+  {
+    this.categoryService.CategoryForm.get("Image").clearValidators();
+    this.categoryService.CategoryForm.get("Image").updateValueAndValidity();
+    this.categoryService.CategoryForm.controls['CategoryID'].setValue(category.CategoryID);
+    this.categoryService.CategoryForm.controls['CategoryName'].setValue(category.CategoryName);
+    this.categoryService.CategoryForm.controls['Description'].setValue(category.Description);
+    this.imageURL=category.Image;
+    //this.categoryService.CategoryForm.controls['Image'].setValue(category.Image);
+  }
+  update(category:Category)
+  {
+    this.categoryService.update(category).subscribe(
+      (data:any)=>
+      {
         this.categories=data;
         this.categoryService.categoryList=data;
+        this.resetForm();
+        console.log("Updated!");
       },
-      (err:any)=>{
-        console.log("Error occurred!");
+      (err:any)=>
+      {
+        console.log("Error!");
       }
     );
+  }
+  resetForm(){
+    this.createForm();
+    this.imageURL=this.defaultImageURL;
   }
 }
